Refactor and add constraint
